+++ bin/hmr.dart
@@ -2,10 +2,6 @@ import 'dart:io';
 
 import 'package:glob/glob.dart';
 import 'package:hmr/hmr.dart';
-import 'package:hmr/src/middlewares/debounce.dart';
-import 'package:hmr/src/middlewares/exclude.dart';
-import 'package:hmr/src/middlewares/ignore.dart';
-import 'package:hmr/src/middlewares/include.dart';
 import 'package:mansion/mansion.dart';
 import 'package:path/path.dart' as path;