+++ test/mocks/app.dart
@@ -1,1 +1,5 @@
-void main() {}
+import 'dart:io';
+
+void main() {
+  exit(0);
+}